pyproject.toml
[project]
name = "cs540-multi-agent"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "anthropic>=0.28",
    "langchain>=0.2",
    "langchain-anthropic>=0.1",
    "langgraph>=0.1",
    "pydantic>=2.0",
    "evalplus>=0.3",
    "datasets>=2.19",
    "pandas>=2.0",
    "scikit-learn>=1.4",
    "python-dotenv>=1.0",
    "jsonlines>=4.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = ["jupyter", "ipykernel", "pytest>=8.0", "pytest-asyncio"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
